requirements.txt
Pillow==8.4.0
Flask==2.0.2